+++ lib/src/iautogram_service.dart
@@ -55,17 +55,17 @@ abstract class IAutogramService {
     DateTime? ifModifiedSince,
   ]);
 
-  /// Device registers itself at the server.
+  /// Registers itself at the server.
   ///
   /// See: <https://generator3.swagger.io/index.html?url=https://autogram.slovensko.digital/openapi.yaml#/Minimal%20Integration/post_devices>
   Future<PostDeviceResponse> registerDevice(PostDeviceRequestBody body);
 
-  /// Device returns list of paired integrations.
+  /// Retrieves a list of paired integrations.
   ///
   /// See: <https://generator3.swagger.io/index.html?url=https://autogram.slovensko.digital/openapi.yaml#/Integration/get_device_integrations>
   Future<GetDeviceIntegrationsResponseBody> listIntegrations();
 
-  /// Device returns list of paired integrations.
+  /// Deletes integration from its paired integrations.
   ///
   /// See: <https://generator3.swagger.io/index.html?url=https://autogram.slovensko.digital/openapi.yaml#/Integration/delete_device_integrations__integration_id_>
   Future<void> deleteIntegration(String integrationId);